requirements.txt
streamlit>=1.32.0
yfinance>=0.2.36
pandas>=2.0.0
plotly>=5.18.0
sqlalchemy>=2.0.0
anthropic>=0.25.0
python-dotenv>=1.0.0
numpy>=1.24.0
